Update spec tests to be agent/master
this is necessary for the pe_mco_shell_agent module since it requires
the puppet_enterprise::profile::mcollective::agent class. This class
isn't installed agent-only and is non-trivial to fetch from github or
another source, so the easiest (and most complete) way to test is to
just use an agent/master configuration.