Add EC2 Beaker nodeset in prep for Travis-CI
This commit adds a working EC2 nodeset file as well as a configuration
script which can be used to pre-configure the Travis-CI node to run
beaker. Mostly this just means creating the .fog file with values from
the environment and generating an ssh key.

Work yet to be done is documenting which environment variables need to
be present for a CI run to complete successfully.