Remove Gemfile.lock
There isn't really a need for a Gemfile.lock in this project. The
Gemfile doesn't matter at all when "deploying" this code, it's only used
during development. When the lockfile is committed, that's just an
opportunity for the versions specified to start to get old, have
security vulnerabilities discovered against them, and eventually get
warnings in Github.

This commit removes the lockfile so security concerns should only be
raised against the Gemfile itself, which is more appropriate for this
kind of project.